Enhance #find_last to handle models backed by UUIDs
The #find_last method in CucumberFactory::Switcher ordered by *id to find the last record. That did not work with models backed with an UUID, or more generally, with models where you couldn't be sure that the *id column was in a certain order. Now #find_last orders by *created_at if available, otherwise it falls back to *id.